feat: support text and EuroOffice document editors
